# services/export-api/openapi.yaml
openapi: '3.1.0'
info:
  title: HoloScript Export API
  version: 1.0.0
  description: |
    REST API for HoloScript compilation and export with SOC 2 compliance foundations.
    All compilations are async (ADR-002). API keys are hashed (ADR-003).
    Audit logs are append-only with integrity chains (ADR-004).
  contact:
    name: Brian X Base Team
    url: https://github.com/brianonbased-dev/Holoscript
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3100
    description: Local development
  - url: https://api.holoscript.dev
    description: Production
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /api/v1/compile:
    post:
      tags: [Compile]
      summary: Submit a compilation job
      description: |
        Submits HoloScript source for async compilation to the specified target.
        Returns 202 Accepted with a job ID for polling (ADR-002).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompileRequest'
      responses:
        '202':
          description: Compilation job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompileJobCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
  /api/v1/compile/{jobId}:
    get:
      tags: [Compile]
      summary: Get compilation job status
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Job status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompileJobStatus'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/compile/{jobId}/output:
    get:
      tags: [Compile]
      summary: Download compilation output
      description: Returns a signed download URL for the compiled output (ADR-005).
      parameters:
        - name: jobId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Download information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompileOutput'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Job not completed
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    CompileRequest:
      type: object
      required: [source, target]
      properties:
        source:
          type: string
          minLength: 1
          maxLength: 1000000
          description: HoloScript source code
        target:
          type: string
          description: Export target (e.g., unity, r3f, usd)
        options:
          type: object
          description: Compilation options
        idempotencyKey:
          type: string
          maxLength: 128
          description: Idempotency key to prevent duplicate submissions
    CompileJobCreated:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending]
        target:
          type: string
        createdAt:
          type: string
          format: date-time
        pollUrl:
          type: string
        message:
          type: string
    CompileJobStatus:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed, cancelled]
        target:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        outputUrl:
          type: string
        outputSizeBytes:
          type: integer
        durationMs:
          type: integer
        error:
          type: string
    CompileOutput:
      type: object
      properties:
        jobId:
          type: string
          format: uuid
        downloadUrl:
          type: string
        contentType:
          type: string
        sizeBytes:
          type: integer
        expiresIn:
          type: integer
          description: URL validity in seconds
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
              details:
                type: array
                items:
                  type: object
    NotFound:
      description: Resource not found (ADR-006)
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Not found
    RateLimited:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
              message:
                type: string
              retryAfterMs:
                type: integer